requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
distlib==0.3.7
docutils @ file:///private/tmp/docutils-20230831-5903-jytmsf/docutils-0.20.1
filelock==3.12.4
fsspec==2023.9.2
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2
numba==0.58.1
numpy==1.26.1
openai-whisper @ git+https://github.com/openai/whisper.git@b38a1f20f4b23f3f3099af2c3e0ca95627276ddf
platformdirs==3.11.0
protobuf==4.21.12
pycparser==2.21
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sympy==1.12
tiktoken==0.3.3
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
whisper~=1.1.10
openai~=0.28.1
Pyrogram~=2.0.106